+++ CHANGELOG.md
@@ -9,6 +9,7 @@ IMPROVEMENTS:
 
 * `vcd_vapp_vm` `disk` has new attribute `size_in_mb` [#433]
 * `resource/vcd_org` and `datasource/vcd_org` include a section `vapp_lease` and a section `vapp_template_lease` to define lease related parameters of depending entities - [GH-432]
+* `datasource/*` - all data sources will return an error when object is not found [GH-xx]
 
 BUG FIXES:
 Fix `vcd_vapp_vm` resource read - independent disks where losing `bus_number` and `unit_number` values after refresh. [#433]